+++ decidim-proposals/spec/forms/decidim/proposals/admin_proposal_form_spec.rb
@@ -7,6 +7,7 @@ module Decidim
     module Admin
       describe ProposalForm do
         it_behaves_like "a proposal form"
+        it_behaves_like "a proposal form with meeting as author"
       end
     end
   end